Adding Csrf Token for delete function

+++ src/AppBundle/Controller/ShowController.php
@@ -3,6 +3,7 @@
 namespace AppBundle\Controller;
 
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use AppBundle\Type\ShowType;
 use AppBundle\Entity\Show;
 use AppBundle\File\FileUploader;
@@ -10,7 +11,8 @@ use AppBundle\Repository\ShowRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
-
+use Symfony\Component\Security\Csrf\CsrfToken;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 
 /**
  * @Route(name="show_")
@@ -34,6 +36,7 @@ class ShowController extends Controller
 
     /**
      * @Route("/search", name="search")
+     * @Method({"POST"})
      */
     public function searchAction(Request $request)
     {
@@ -111,19 +114,26 @@ class ShowController extends Controller
     }
 
     /**
-     *  @Route("/delete", name="delete")
+     * @Route("/delete", name="delete")
+     * @Method({"POST"})
      */
-    public function deleteAction(Request $request)
+    public function deleteAction(Request $request, CsrfTokenManagerInterface $csrfTokenManager)
     {
         $showId = $request->request->get('show_id');
         $show = $this->getDoctrine()->getRepository('AppBundle:Show')->findOneBy(['id' => $showId]);
         if(!$show)
             throw new NotFoundHttpException('There is no show with the id %d', $showId);
-        if(strlen($show->getMainPicture()) > 0)
-            unlink($this->getParameter('kernel.project_dir').'/web'.$this->getParameter('upload_directory_file').'/'.$show->getMainPicture());
-        $doctrine->getManager()->remove($show);
-        $doctrine->getManager()->flush();
-        $this->addFlash('success', 'The show has been successfully deleted');
+        $csrfToken = new CsrfToken('delete_show', $request->request->get('_csrf_token'));
+        if($csrfTokenManager->isTokenValid($csrfToken)){
+            if(strlen($show->getMainPicture()) > 0)
+                unlink($this->getParameter('kernel.project_dir').'/web'.$this->getParameter('upload_directory_file').'/'.$show->getMainPicture());
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($show);
+            $em->flush();
+            $this->addFlash('success', 'The show has been successfully deleted');
+        }else{
+            $this->addFlash('danger', 'The csrf token is not valid. The detection was not completed.');
+        }
         return $this->redirectToRoute('show_list');
     }
 }
